Crate sdk_utils

Utilities crate shared across all SDK crates

Modules

chain_spec

Subspace chain spec related utilities

Macros

generate_builder

Useful macro to generate some common methods and trait implementations for builders

Structs

ByteSize

Container for number of bytes.

DestructorSet

A General purpose set of destructors consist of sync destructor, async destructor and normal object it invokes destructors and destroy normal in reverse order

Multiaddr

Multiaddr is a wrapper around libp2p one

MultiaddrWithPeerId

Multiaddr with peer id

PublicKey

Public key type

Rpc

Rpc implementation over jsonrpsee_core debug rpc module

StorageKey

Substrate storage key abstraction

Enums

Ss58ParsingError

An error type for SS58 decoding.

TaskOutput

Output that indicates whether the task was cancelled or successfully completed

Functions

is_default

Useful predicate for serde, which allows to skip type during serialization

task_spawn

Spawn task with provided name (if possible)

task_spawn_blocking

Spawn task with provided name (if possible)